Create greetings.py
I can't seem to get multiple instances of 'if' to run accordingly. Break doesn't work but without break, python just cycles from one if to the next. Any ideas? do they both need to be in a loop to be tested for and then exited? I just want the user's choice of goodbye or hello to be the only thing that runs before the program stops.