Create basic_calc.py
I couldn't get the last segment to run (though I didn't try particularly hard to learn how to make it run) which would have given the user the chance to exit the program if desired. The addition of checking for the run variable being true was enough to cause the program to loop and enable multiple calculations. Kinda fun. I had to look up how to concatenate int's onto the end of strings and also the error handling (which I don't think functions correctly either) but otherwise this was from memory. This was a test for me as I have been doing a lot of Swift programming and wanted to see how much Python I remembered.