Create port_scanner.py
This code came from a short Youtube tutorial on creating a basic port scanner. When run from the Python interpreter it never returns any value though it behaves like it is scanning; a network traffic monitor will echo that it is, in fact, passing traffic. But the result is never displayed and the program must be overridden to complete. The YT video shows the program being run from a command line and is being passed an IP address even though the program asks the user to input one during run time. I provided the link to the YT video in case anyone wants to see the original code or get the link to the descriptive website. If anyone knows why this is never resolving ports and displaying in the console, please let me know.