Create factorial.py
This is from codingame.com. It's a listing under recursion for Python factorials. An algorithm. The int exchange in the beginning input is required to allow the user to enter data. For some reason, assigning the value of user input to a variable creates a string where this algorithm needs an int.